backend/requirements.txt
asgiref
autopep8
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
pycodestyle
PyJWT
pytz
sqlparse
toml
tzdata
requests
mistralai